Fase 3.39: npoi-demo passes — Dictionary ordering, interface overload resolution, real System.Reflection
Completes the WIP from c148242: examples/npoi-demo now opens a real
legacy .xls, constructs a real HSSFWorkbook, and prints real cell data
deterministically across runs.

Root cause of NotOLE2FileException: nativeDict had no memory of
insertion order, so Dictionary/Hashtable enumeration got Go's own
randomized map order instead of the stable order real code (NPOI's own
FileMagicContainer.ValueOf) silently depends on. Fixed with an
insertion-ordered `order` slice threaded through every read/write path.

Also fixed, all found running the real file through the interpreter:
- Overload resolution never checked a candidate's implemented
  interfaces, only its base-class chain — AreaPtg(ILittleEndianInput)
  vs AreaPtg(AreaReference), a same-arity pair, picked the wrong one
  for a stream argument.
- bcl.NativeTypeName reported one fixed name per Go struct regardless
  of which real type constructed it (Hashtable misreported as
  Dictionary`2, ArrayList as List`1), corrupting virtual-dispatch
  retries.
- A real System.Reflection subsystem: Type.GetConstructor/GetMethod/
  GetField plus ConstructorInfo/MethodInfo/FieldInfo's own Invoke/
  GetValue, needed by RecordFactory's own reflection-based type
  registry (not Reflection.Emit — no code generation).
- Literal/const static fields never got their real compile-time value
  (decoded via the ECMA-335 Constant table now).
- String.Concat's single-array-argument overload wasn't unwrapped.
- New: System.Threading.Monitor, System.IO.Path, Comparer`1,
  SortedList, legacy Stack, StringBuilder.Insert (backing store
  switched from strings.Builder to a plain string), Encoding.Unicode/
  BigEndianUnicode as real UTF-16 (not a UTF-8-passthrough alias).

Known remaining limitation: formula-text rendering shows column
letters as their numeric code points (vmnet has no distinct char
Kind) — an architecture change, out of scope here.

NPOI@2.8.0 netstandard-lite: 97.0% -> 97.3% clean, as a side effect.

+++ examples/npoi-demo/main.go
@@ -84,7 +84,7 @@ func main() {
 		if err != nil {
 			log.Fatalf("get_LastCellNum: %v", err)
 		}
-		cellCount := int(lastCell.Native().(int16))
+		cellCount := int(lastCell.Native().(int32))
 
 		for c := 0; c < cellCount; c++ {
 			cell, err := rowInst.Call("GetCell", vmnet.Int32(int32(c)))
